Service/Backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
motor==3.4.0
pymongo==4.8.0
pydantic[email]==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0
openpyxl==3.1.5